Refactor logging and request ID generation into shared utility functions
